partialize, sharedizize, zazezizize
