+++ Content.Server/AlternateDimension/AlternateDimensionSystem.Portal.cs
@@ -1,0 +1,9 @@
+namespace Content.Server.AlternateDimension;
+
+public sealed partial class AlternateDimensionSystem
+{
+    private void InitializePortal()
+    {
+
+    }
+}
